Add validation and unlock handlers to purchase service
The validation handler is added so that the app can handle its own receipt validation to whatever degree desired and return a boolean of validity. The unlock features handler allows the app to unlock any purchased features and confirm success by returning a boolean so that the purchase won't be completed in the App Store without the user getting the goods.