Add validation and unlock handler step to restore purchases
This is to make restoring purchases similar to the buying step.
The restore function will now use the same validation handler
and unlock handler the app registers for the buy step, so that
the app will verify the receipt and unlock whatever is necessary.
However, unlike the buy function, it does not update observers.